Valerio/AcqToInfluxDB/requirements.txt
pyserial
matplotlib
influxdb
